extract sample activity to rx package
