+++ sample/src/main/java/ru/tinkoff/tirecycler/rx/SampleTiRecyclerHolderFactory.kt
@@ -1,8 +1,9 @@
-package ru.tinkoff.tirecycler
+package ru.tinkoff.tirecycler.rx
 
 import android.view.View
 import ru.tinkoff.mobile.tech.ti_recycler.base.BaseViewHolder
 import ru.tinkoff.mobile.tech.ti_recycler_rx2.base.RxHolderFactory
+import ru.tinkoff.tirecycler.R
 import ru.tinkoff.tirecycler.items.HeaderUiViewHolder
 import ru.tinkoff.tirecycler.items.MultiChoiceCheckViewHolder
 import ru.tinkoff.tirecycler.items.TextUiViewHolder
